+++ services/src/main/java/ar/edu/itba/paw/services/SubscriptionServiceImpl.java
@@ -42,7 +42,7 @@ public class SubscriptionServiceImpl implements SubscriptionService {
     @Override
     @Transactional(readOnly = true)
     public PageModel<Users> listSubscriptions(final int subscriberId, final int page, final int pageSize) {
-        final int totalItems = subscriptionDao.countSubscriptions(subscriberId);
+        final long totalItems = subscriptionDao.countSubscriptions(subscriberId);
         return new PageModel<>(
                 subscriptionDao.listSubscriptions(subscriberId, page, pageSize), page, pageSize, totalItems);
     }
